WIP refactor MainWindow into child WPF Pages
