+++ wv2util/Pages/IReloadable.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace wv2util
+{
+    public interface IReloadable
+    {
+        void Reload();
+        bool Reloading { get; }
+
+        event EventHandler ReloadingChanged;
+    }
+}
