+++ wv2util/Pages/RuntimesPage.xaml
@@ -1,0 +1,63 @@
+﻿<Page x:Class="wv2util.RuntimesPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:wv2util"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="RuntimesPage">
+    <Page.Resources>
+        <ObjectDataProvider x:Key="RuntimeList" ObjectType="{x:Type local:AppState}" MethodName="GetRuntimeList"/>
+    </Page.Resources>
+
+    <Grid DataContext="{Binding Source={StaticResource RuntimeList}}">
+        <Grid.RowDefinitions>
+            <RowDefinition/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+
+        <ListView x:Name="RuntimeList" ItemsSource="{Binding}" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Grid.Row="0">
+            <ListView.View>
+                <GridView>
+                    <GridViewColumn DisplayMemberBinding="{Binding RuntimeLocation}">
+                        <GridViewColumnHeader Click="GridViewColumnHeader_Runtime_Path_Click">Path</GridViewColumnHeader>
+                    </GridViewColumn>
+                    <GridViewColumn DisplayMemberBinding="{Binding Version}">
+                        <GridViewColumnHeader Click="GridViewColumnHeader_Runtime_Version_Click">Version</GridViewColumnHeader>
+                    </GridViewColumn>
+                    <GridViewColumn DisplayMemberBinding="{Binding Channel}">
+                        <GridViewColumnHeader Click="GridViewColumnHeader_Runtime_Channel_Click">Channel</GridViewColumnHeader>
+                    </GridViewColumn>
+                </GridView>
+            </ListView.View>
+        </ListView>
+        <Expander Grid.Row="1">
+            <Expander.Header>
+                <Label Content="Find more WebView2 Runtimes" />
+            </Expander.Header>
+            <TextBlock HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="30,0,0,10">
+              <TextBlock>
+                <Hyperlink NavigateUri="https://docs.microsoft.com/en-us/deployedge/microsoft-edge-release-schedule" RequestNavigate="Hyperlink_RequestNavigate">Microsoft Edge release schedule</Hyperlink>
+              </TextBlock>
+              <LineBreak/>
+              <LineBreak/>
+              <TextBlock>
+                <Hyperlink NavigateUri="https://developer.microsoft.com/en-us/microsoft-edge/webview2/#:~:text=Evergreen%20Standalone%20Installer" RequestNavigate="Hyperlink_RequestNavigate">WebView2 Runtime</Hyperlink> (<Hyperlink NavigateUri="https://docs.microsoft.com/en-us/deployedge/microsoft-edge-relnote-stable-channel" RequestNavigate="Hyperlink_RequestNavigate">browser release notes</Hyperlink>, <Hyperlink NavigateUri="https://docs.microsoft.com/en-us/microsoft-edge/webview2/release-notes" RequestNavigate="Hyperlink_RequestNavigate">SDK release notes</Hyperlink>)
+              </TextBlock>
+              <LineBreak/>
+              <TextBlock>
+                <Hyperlink NavigateUri="https://www.microsoft.com/en-us/edge/download/insider?form=MA13FJ" RequestNavigate="Hyperlink_RequestNavigate">Microsoft Canary, Dev and Beta Channel</Hyperlink> (<Hyperlink NavigateUri="https://learn.microsoft.com/en-us/deployedge/microsoft-edge-relnote-beta-channel" RequestNavigate="Hyperlink_RequestNavigate">beta release notes</Hyperlink>)
+              </TextBlock>
+              <LineBreak/>
+              <TextBlock>
+                <Hyperlink NavigateUri="https://developer.microsoft.com/en-us/microsoft-edge/webview2/#:~:text=ARM64-,Fixed%20Version,-Select%20and%20package" RequestNavigate="Hyperlink_RequestNavigate">WebView2 Runtime Fixed Version</Hyperlink>
+              </TextBlock>
+              <LineBreak/>
+              <TextBlock>
+                <Hyperlink NavigateUri="https://www.microsoft.com/en-us/edge/business/download#:~:text=Need%20to%20download%20older%20builds%20and%20versions%3F" RequestNavigate="Hyperlink_RequestNavigate">Microsoft Edge older versions</Hyperlink>
+              </TextBlock>
+            </TextBlock>
+        </Expander>
+    </Grid>
+</Page>
